Moved more scripts to buildSrc and refactored buildSrc
- Moved wrapper.gradle.kts to buildSrc
- Moved resumeBuild.gradle to buildSrc
- Moved minifiedDependencies.gradle to buildSrc
- Moved dependenciesCheck.gradle to buildSrc
- Moved updateReleasedVersions.gradle to buildSrc
- Introduced configuration submodule for developer focused configuration
- Created buildSrc submodule for buildquality
- Created buildSrc submodule for packaging
- Created buildSrc submodule for profiling
- Created buildSrc submodule for versioning
- Introduced org.gradle.gradlebuild naming convention
- Introduced project-groups file for providing groups of subprojects
