+++ buildSrc/settings.gradle.kts
@@ -25,12 +25,18 @@ fun String.toKebabCase() =
 
 rootProject.name = "buildSrc"
 
+// Please preserve alphabetical order
 include("binaryCompatibility")
 include("build")
-include("plugins")
-include("kotlinDsl")
+include("buildquality")
 include("cleanup")
+include("configuration")
+include("kotlinDsl")
+include("packaging")
+include("plugins")
+include("profiling")
 include("testing")
+include("versioning")
 
 fun buildFileNameFor(projectDirName: String) =
     "$projectDirName.gradle.kts"
