+++ buildSrc/subprojects/configuration/configuration.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    `java-gradle-plugin`
+}
+
+apply { plugin("org.gradle.kotlin.kotlin-dsl") }
+
+dependencies {
+    compile(project(":kotlinDsl"))
+}
