+++ buildSrc/subprojects/configuration/src/main/kotlin/org/gradle/gradlebuild/build-environment.kt
@@ -1,0 +1,5 @@
+package org.gradle.gradlebuild
+
+object BuildEnvironment {
+    val isCiServer = "CI" in System.getenv()
+}
