+++ buildSrc/subprojects/configuration/src/main/kotlin/org/gradle/gradlebuild/project-groups.kt
@@ -1,0 +1,17 @@
+package org.gradle.gradlebuild
+
+object ProjectGroups {
+    val excludedFromVulnerabilityCheck = setOf(
+        ":buildScanPerformance",
+        ":distributions",
+        ":docs",
+        ":integTest",
+        ":internalAndroidPerformanceTesting",
+        ":internalIntegTesting",
+        ":internalPerformanceTesting",
+        ":internalTesting",
+        ":performance",
+        ":runtimeApiInfo",
+        ":smokeTest",
+        ":soak")
+}
