--- buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/BuildEnvironment.kt
@@ -1,5 +1,0 @@
-package org.gradle
-
-object BuildEnvironment {
-    val isCiServer = "CI" in System.getenv()
-}
