+++ buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/gradlebuild/tools/ResumeBuildPlugin.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.gradlebuild.tools
+
+import org.gradle.api.GradleException
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class ResumeBuildPlugin: Plugin<Project> {
+    override fun apply(project: Project): Unit = project.run {
+        val resumeTask = project.property("resume")
+
+        if (resumeTask != null) {
+            gradle.taskGraph.whenReady {
+                val allTasks = allTasks
+                    val resumeIndex = allTasks.indexOfFirst { it.path == resumeTask }
+                if (resumeIndex < 0) throw GradleException("Can't resume from $resumeTask because no such task is scheduled for execution")
+                allTasks.subList(0, resumeIndex).forEach { it.enabled = false }
+            }
+        }
+    }
+}
