--- buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/plugins/codequality/FailDependencyResolutionAtConfiguration.kt
@@ -1,44 +1,0 @@
-/*
- * Copyright 2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.gradle.plugins.codequality
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-
-
-open class FailDependencyResolutionAtConfiguration : Plugin<Project> {
-
-    override fun apply(project: Project): Unit = project.run {
-        var projectsEvaluated = false
-
-        gradle.projectsEvaluated {
-            projectsEvaluated = true
-        }
-
-        allprojects {
-            configurations.all {
-                val configName = this.name
-                incoming.beforeResolve {
-                    if (!projectsEvaluated) {
-                        throw Exception("Configuration $configName of project ${this@allprojects.project.name} is being resolved at configuration time.")
-                    }
-                }
-            }
-        }
-
-    }
-}
-
