+++ buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/plugins/reporting/CiReportingPlugin.kt
@@ -1,7 +1,7 @@
 package org.gradle.plugins.reporting
 
 import me.champeau.gradle.japicmp.JapicmpTask
-import org.gradle.BuildEnvironment
+import org.gradle.gradlebuild.BuildEnvironment
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.Task
