+++ buildSrc/subprojects/profiling/src/main/kotlin/org/gradle/gradlebuild/profiling/JmhPlugin.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.gradlebuild.profiling
+
+import me.champeau.gradle.JMHPluginExtension
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+import org.gradle.kotlin.dsl.*
+
+open class JmhPlugin : Plugin<Project> {
+
+    override fun apply(project: Project): Unit = project.run {
+        apply {
+            plugin("me.champeau.gradle.jmh")
+        }
+
+        configurations.getByName("jmhImplementation").extendsFrom(configurations["implementation"])
+
+        configure<JMHPluginExtension> {
+            isIncludeTests = false
+            resultFormat = "CSV"
+        }
+    }
+}
+
+
