--- gradle/dependenciesCheck.gradle
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-buildscript {
-    repositories {
-        maven {
-            url "https://plugins.gradle.org/m2/"
-        }
-    }
-    dependencies {
-        classpath 'org.owasp:dependency-check-gradle:3.1.0'
-    }
-}
-
-configure(projectsIncludedInDistribution()) {
-    apply plugin: org.owasp.dependencycheck.gradle.DependencyCheckPlugin
-
-    dependencyCheck {
-        failBuildOnCVSS = 8 // 10 is the maximum
-        skipConfigurations = ['jmh']
-    }
-}
-
-def projectsIncludedInDistribution() {
-    def excluded = [':buildScanPerformance', ':distributions', ':docs', ':integTest', ':internalAndroidPerformanceTesting', ':internalIntegTesting', ':internalPerformanceTesting', ':internalTesting', ':performance', ':runtimeApiInfo', ':smokeTest', ':soak'] as Set<String>
-    subprojects.findAll { prj -> !excluded.contains(prj.path) }
-}
