--- gradle/jmh.gradle
@@ -1,38 +1,0 @@
-/*
- * Copyright 2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-buildscript {
-    repositories {
-        maven {
-            url "https://plugins.gradle.org/m2/"
-        }
-    }
-    dependencies {
-        classpath 'me.champeau.gradle:jmh-gradle-plugin:0.4.5'
-    }
-}
-
-apply plugin: me.champeau.gradle.JMHPlugin
-
-configurations {
-    jmhImplementation.extendsFrom(implementation)
-}
-
-jmh {
-    includeTests = false
-    resultFormat = 'CSV'
-}
-
