--- gradle/resumeBuild.gradle
@@ -1,10 +1,0 @@
-def resumeTask = project.property("resume")
-
-if (resumeTask) {
-    gradle.taskGraph.whenReady { graph ->
-        def allTasks = graph.allTasks
-        def resumeIndex = allTasks.findIndexOf { it.path == resumeTask }
-        if (resumeIndex < 0) throw new GradleException("Can't resume from $resumeTask because no such task is scheduled for execution")
-        allTasks.subList(0, resumeIndex)*.enabled = false
-    }
-}
